Improve performance of wrong cache misses
Some cache misses are 'wrong', and are due to some inputs not having
been restored from cache before the target is being looked up. This
change improves performance in that case by maintaining a pointer to the
'latest' execution log, and using it when the cache misses to try to
guess which inputs need to be built.